Enhance semester management features
- Introduced API endpoint for updating semesters via the new `apiUpdate` method in `SemesterController`.
- Added `ApiUpdateSemesterRequest` for request validation during semester updates.
- Updated frontend components to support the new API for semester updates, including form handling and error management.
- Refactored existing routes to include the new API endpoint for semester updates.
- Improved documentation to reflect the new update capabilities for semesters.

+++ app/Models/Semester.php
@@ -242,7 +242,7 @@ class Semester extends Model
     {
         $now = Carbon::now();
 
-        return static::where('is_active', true)
+        return (int)static::where('is_active', true)
             ->where('end_date', '<', $now)
             ->update(['is_active' => false]);
     }
